PEAR-1960 - move migration to the right place
